refactor plivo xml module and make it available as both plivo.xml and plivoxml
