+++ .gitignore
@@ -8,3 +8,4 @@ dist/
 *.pyo
 ./tests/htmlcov/
 .tox/
+venv/
