+++ plivo/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
-from . import exceptions, plivoxml
+from . import exceptions
 from .rest import Client as RestClient
+import xml as plivoxml
