+++ plivo/xml/xmlUtils.py
@@ -1,0 +1,6 @@
+import six
+
+def map_type(val):
+    if isinstance(val, bool):
+        return six.text_type(val).lower()
+    return six.text_type(val)
